# Optimize Struct's accessors