@maximecb 2021-11-18
@XrXr @jhawthorn can you comment on whether this PR increases or decrease the complexity in YJIT, and how easy/hard it would be for us to support this change?