@jhawthorn 2021-11-18
> can you comment on whether this PR increases or decrease the complexity in YJIT, and how easy/hard it would be for us to support this change?

I think it would very slightly increase complexity, since it's adding two new method types, but those method types should be pretty easy to implement. We should also be able to generate better code for these than we currently do (can now read from the memory location instead of calling the old builtin function).