@XrXr 2021-11-18
Yeah, we would need to add code to YJIT if we want to support this new method type. The actual logic is similar to getting an ivar, but a bit simpler because structs have a fixed layout. The relevant code is here:
https://github.com/ruby/ruby/blob/ab737b19197c63b84dad9944045a2fd2dc369264/internal/struct.h#L121-L144

So it's a flag test to find the buffer, and then reading at a fixed index.