@maximecb 2021-11-18
Since you guys don't seem concerned about the complexity, we can approve it. Would either of you want to implement the new struct access mechanism?