@schneems 2021-11-19
After compiling with these changes I get a warning:

```
struct.c:267:1: warning: unused function 'struct_pos_num' [-Wunused-function]
struct_pos_num(VALUE s, VALUE idx)
^
1 warning generated.
```