@paracycle 2021-12-08
This PR seems to have changed the public API of `Struct` member setter methods.

Before:
```ruby
S = Struct.new(:foo)
S.instance_method(:foo=).parameters
#=> [[:req, :_]]
```

After:
```ruby
S = Struct.new(:foo)
S.instance_method(:foo=).parameters
#=> [[:rest]]
```

Is this intentional? I find it weird that a setter method looks like it is accepting zero or more arguments instead of a single required argument.